Enables Youtube/web full screen with a clean UX
After trying many solutions, I managed to stich
this one together which uses a custom web view and
overrides onShowCustomView and onHideCustomView which
in turn implements hiding and showing a full screen video.
Refactored some code in the `WatchContentFragment`.

Alternate soultions:
1) Youtube Video Player API: obvious choice but costs $

2) android-youtube-player:
https://github.com/PierfrancescoSoffritti/android-youtube-player
(Allows for ALOT of customizability, however seems
like an overkill for now, and would also require
much more effort to even get a simple webview working
like the iFrame. Simple solution of showing full screen
would just be to show a button outside of the webview. not the best ux.)

Sources:
- https://stackoverflow.com/questions/15768837/playing-html5-video-on-fullscreen-in-android-webview/16179544#16179544
- https://github.com/akhgupta/WebviewVideo